pyproject.toml
[project]
name = "ozon-mcp-server"
version = "0.1.0"
description = "The first MCP server for Ozon Seller API — Russia's largest e-commerce marketplace"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [{name = "Azat KH"}]
keywords = ["mcp", "ozon", "marketplace", "ai", "llm", "fastmcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "fastmcp>=2.0",
    "httpx>=0.27",
    "pydantic>=2.11",
    "pydantic-settings>=2.7",
    "asyncpg>=0.30",
    "redis>=5.0",
    "tenacity>=9.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "pytest-httpx>=0.35",
    "ruff>=0.8",
    "mypy>=1.13",
    "bandit>=1.8",
    "pip-audit>=2.7",
    "safety>=3.0",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ozon_mcp_server"]

[tool.ruff]
target-version = "py312"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM", "TCH", "PTH"]
ignore = ["S101"]  # assert in tests

[tool.ruff.lint.isort]
known-first-party = ["ozon_mcp_server"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src/ozon_mcp_server --cov-report=term-missing --cov-report=xml"

[tool.bandit]
targets = ["src"]
exclude_dirs = ["tests"]

[tool.coverage.run]
source = ["src/ozon_mcp_server"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
